+++ tests/clients/test_opening_explorer.py
@@ -1,7 +1,12 @@
 import pytest
 import requests_mock
-
-from berserk import Client, OpeningStatistic
+from berserk import (
+    Client,
+    OpeningStatistic,
+    MastersOpeningStatistic,
+    PlayerOpeningStatistic,
+)
+from typing import List
 
 from utils import validate, skip_if_older_3_dot_10
 
@@ -39,14 +44,14 @@ class TestLichessGames:
 
 
 class TestMasterGames:
+    @skip_if_older_3_dot_10
     @pytest.mark.vcr
     def test_result(self):
+        """Verify that the response matches the typed-dict"""
         res = Client().opening_explorer.get_masters_games(
             play=["d2d4", "d7d5", "c2c4", "c7c6", "c4d5"]
         )
-        assert res["white"] == 1667
-        assert res["black"] == 1300
-        assert res["draws"] == 4428
+        validate(MastersOpeningStatistic, res)
 
     @pytest.mark.vcr
     def test_export(self):
@@ -75,9 +80,7 @@ class TestPlayerGames:
         result = Client().opening_explorer.get_player_games(
             player="evachesss", color="white", wait_for_indexing=True
         )
-        assert result["white"] == 125
-        assert result["draws"] == 18
-        assert result["black"] == 133
+        validate(PlayerOpeningStatistic, result)
 
     @pytest.mark.vcr
     @pytest.mark.default_cassette("TestPlayerGames.results.yaml")
@@ -87,34 +90,15 @@ class TestPlayerGames:
             color="white",
             wait_for_indexing=False,
         )
-        assert result == {
-            "white": 0,
-            "draws": 0,
-            "black": 0,
-            "moves": [],
-            "recentGames": [],
-            "opening": None,
-            "queuePosition": 0,
-        }
+        validate(PlayerOpeningStatistic, result)
 
     @pytest.mark.vcr
     @pytest.mark.default_cassette("TestPlayerGames.results.yaml")
     def test_stream(self):
-        result = list(
-            Client().opening_explorer.stream_player_games(
-                player="evachesss",
-                color="white",
-            )
+        iterator = Client().opening_explorer.stream_player_games(
+            player="evachesss",
+            color="white",
         )
-        assert result[0] == {
-            "white": 0,
-            "draws": 0,
-            "black": 0,
-            "moves": [],
-            "recentGames": [],
-            "opening": None,
-            "queuePosition": 0,
-        }
-        assert result[-1]["white"] == 125
-        assert result[-1]["draws"] == 18
-        assert result[-1]["black"] == 133
+        # Just test that the stream yields at least one result
+        result = next(iterator)
+        validate(PlayerOpeningStatistic, result)